use managers to build references and move bep logic to bephelper
